Rewrite to lua and remove all things I don't want/need
I had some problems with this plugin together `unnest.nvim`. That could
be fixed by configuring the plugin, but then I also noticed that I had
some problems with the details window when using extui. That couldn't be
fixed without modifying the source code.

It doesn't look like this is too active, so I'll just rewrite this to
lua and maintain it myself.

I am not interested in adding new stuff to this or configurations. I
want to keep it extremely simple